# spec/swagger.yaml
definitions:
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.synthesizeRequest:
    properties:
      text:
        type: string
    type: object
  handlers.synthesizeResponse:
    properties:
      content_type:
        type: string
      size:
        type: integer
      temp_id:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/tts/synthesize:
    post:
      consumes:
      - application/json
      description: Stream-synthesize text using the bot's configured TTS model, write
        to temp file
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Text to synthesize
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.synthesizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.synthesizeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Synthesize speech for a bot
      tags:
      - bots
...
swagger: "2.0"